比较重要的进展：实现zip（图片包）的展现和删除！（直接从user_dashboard点击打开按钮即可）

+++ Orbital_django/file_viewer/models.py
@@ -3,6 +3,7 @@ from django.db import models
 from home import models as home_models
 from django.dispatch import receiver
 import os
+import shutil
 
 
 def upload_to(instance, filename):
@@ -26,6 +27,9 @@ def delete_local_file(sender, instance, **kwargs):
     file_location = instance.file_field.name
 
     file_local_location = instance.file_field.storage.path(instance.file_field)
+    file_dirname, file_name_and_extension = os.path.split(file_local_location)
+    file_name, extension = file_name_and_extension.split(".")
+    associated_folder_location = os.path.join(file_dirname, file_name)
     type_level_dir_name = os.path.dirname(file_local_location)
 
     # instance.file_field.storage will return an instance of Storage' subclass,
@@ -34,16 +38,20 @@ def delete_local_file(sender, instance, **kwargs):
     # associated with the corresponding file_field of the model to be deleted
     instance.file_field.storage.delete(file_location)
 
+    # delete associate image folder
+    if os.path.isdir(associated_folder_location):
+        shutil.rmtree(associated_folder_location)
+
     # if the type-level folder such as folder named "doc" or "pdf" is empty
     # that is, if the user does not have documents of this type
     # then delete this type_level folder
-    if os.listdir(type_level_dir_name).__len__() == 0:
+    if os.path.isdir(type_level_dir_name) and os.listdir(type_level_dir_name).__len__() == 0:
         user_level_dir_name = os.path.dirname(type_level_dir_name)
         os.rmdir(type_level_dir_name)
         # if the user-level folder is empty
         # that is, if the user does not have any documents
         # then delete this user_level folder
-        if os.listdir(user_level_dir_name).__len__() == 0:
+        if os.path.isdir(user_level_dir_name) and os.listdir(user_level_dir_name).__len__() == 0:
             os.rmdir(user_level_dir_name)
